add CRF as postprocessing
