+++ farmer/ncc/__init__.py
@@ -9,6 +9,7 @@ from .dataset import *
 from .validations import *
 from .predictions import *
 from .preprocessing import *
+from .postprocessing import *
 from .tasks import *
 from .schedulers import *
 from .losses import *
