+++ farmer/ncc/postprocessing/__init__.py
@@ -1,0 +1,1 @@
+from .postprocessing import do_crf
